test(serviceowner): consolidate bulk set validator tests
